requirements.txt
matplotlib==3.7.2
numpy==1.24.3
opencv_contrib_python==4.8.1.78
pandas==2.0.3
scikit_learn==1.3.0
seaborn==0.13.0
torchsummary==1.5.1
#pip3 install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118